🦄 refactor(base.ts and types.ts):
I split the base.ts file into middleware and environments files to adhere to provide a descriptive name for these functions. I moved from types.d.ts to a types folder to hold the definition files, and created one to match the middleware types.
